# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Klori",
        "license": {
            "name": "Klori"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Api Documentations Klori"
        }
    ],
    "paths": {
        "/web/auth/change": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Restablecer contraseña",
                "operationId": "Change",
                "requestBody": {
                    "description": "Formato de envío de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "t": {
                                        "description": "Token de usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña nueva de usuario",
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "description": "Confirma nueva de usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha enviado el enlace de restablecer contraseña correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChangeResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/web/auth/forgot": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Restablecer contraseña",
                "operationId": "Forgot",
                "requestBody": {
                    "description": "Formato de envío de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Nombre de usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha enviado el enlace de restablecer contraseña correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForgotResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/web/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicio de sesión de usuario",
                "operationId": "Login",
                "requestBody": {
                    "description": "Formato de envío de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Nombre de usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "El usuario a iniciado sesión correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no fue autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ChangeResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ForgotResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResponseData"
                    }
                },
                "type": "object"
            },
            "LoginResponseData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "token": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "admin": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "authorization": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}